+++ Paradigm/Changes.md
@@ -1,7 +1,7 @@
 # Utopia Paradigm - List of Changes
 
 ## v1.7.1 (in development)
-Latest Flow support
+Working on distance vectors
 ### Bugfixes
 - Attempting to calculate a value for a **Change** with duration of zero would previously cause a 
   **StackOverFlowException** in situations where the **Change**'s `.toString` involved such value calculations.
@@ -29,6 +29,7 @@ Latest Flow support
   - Added `.forAbbreviation(String)` and `.from(String)`
 ### Other changes
 - **Distance** now extends **Combinable**
+- **Dimensions** now better utilizes **OptimizedIndexedSeq**
 
 ## v1.7 - 04.10.2024
 This update focuses on the **Polygonic** trait (now named **Polygon**), refactoring many of the existing functions.